% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bactaxid.R
\name{classify}
\alias{classify}
\title{Clasificar genomas usando BacTaxID}
\usage{
classify(
  fasta_files,
  genus = NULL,
  db_path = NULL,
  db_dir = tools::R_user_dir("bactaxidR", which = "data"),
  bin_path = NULL,
  bin_dir = tools::R_user_dir("bactaxidR", which = "data"),
  verbose = TRUE
)
}
\arguments{
\item{fasta_files}{Vector de caracteres con las rutas a los archivos FASTA de los genomas a clasificar.}

\item{genus}{Carácter (opcional). Género bacteriano de la base de datos a usar. Si se proporciona,
se buscará automáticamente `<genus>.db` en `db_dir`. Si no se encuentra, se descargará automáticamente.}

\item{db_path}{Carácter (opcional). Ruta directa al archivo de base de datos DuckDB (`.db`). Si se especifica, tiene prioridad sobre `genus`.}

\item{db_dir}{Carácter. Directorio donde buscar o descargar las bases de datos de referencia.}

\item{bin_path}{Carácter (opcional). Ruta directa al ejecutable de BacTaxID. Si es NULL, se buscará en `bin_dir`.}

\item{bin_dir}{Carácter. Directorio donde buscar o descargar el ejecutable de BacTaxID.}

\item{verbose}{Lógico. Si es TRUE, muestra mensajes informativos del proceso y del ejecutable.}
}
\value{
Un data.frame con los resultados de la clasificación conteniendo columnas como `query_id`, `signature`, `best_hit`, `similarity`, `levels_reached` y códigos taxonómicos.
}
\description{
Realiza la clasificación taxonómica de un conjunto de genomas (en formato FASTA o FASTA comprimido)
contra una base de datos de referencia DuckDB específica para un género.
}
