% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bactaxid.R
\name{download_bactaxid}
\alias{download_bactaxid}
\title{Descargar el ejecutable BacTaxID}
\usage{
download_bactaxid(
  dest_dir = tools::R_user_dir("bactaxidR", which = "data"),
  force = FALSE
)
}
\arguments{
\item{dest_dir}{Ruta al directorio donde se guardará el ejecutable. Por defecto usa el directorio de datos del usuario (`tools::R_user_dir("bactaxidR", which = "data")`).}

\item{force}{Lógico. Si es TRUE, descarga el ejecutable incluso si ya existe localmente.}
}
\value{
Ruta absoluta al ejecutable descargado.
}
\description{
Descarga el ejecutable precompilado de BacTaxID (versión 1.0 para Linux) desde el repositorio oficial de GitHub.
}
