% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_db.R
\name{download_db}
\alias{download_db}
\title{Download reference database from Zenodo}
\usage{
download_db(
  genus,
  dest_dir = tools::R_user_dir("bactaxidR", which = "data"),
  force = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{genus}{Character. The bacterial genus of interest (e.g., "Escherichia", "Salmonella"). Case-insensitive.}

\item{dest_dir}{Path to the directory where the uncompressed database will be saved. Defaults to the user data directory (`tools::R_user_dir("bactaxidR", which = "data")`).}

\item{force}{Logical. If `TRUE`, downloads the database even if it already exists locally.}

\item{verbose}{Logical. If `TRUE`, prints progress messages.}
}
\value{
Absolute path to the decompressed DuckDB database (`.db`).
}
\description{
Downloads and decompresses the DuckDB reference database for a specific bacterial genus from the Zenodo record (17791772).
}
