% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bactaxid.R
\name{get_code_table}
\alias{get_code_table}
\title{Importar la tabla 'code' de la base de datos DuckDB}
\usage{
get_code_table(db_path, full_table = FALSE)
}
\arguments{
\item{db_path}{CarĂ¡cter. Ruta a la base de datos DuckDB (`.db`).}

\item{full_table}{LĂ³gico. Si es `TRUE`, importa la tabla completa (incluyendo `signature` y las columnas `_state`). Si es `FALSE` (por defecto), solo importa las columnas `sample` y las columnas de Ă­ndices (`_int`) y cĂ³digos completos (`_full`) de cada nivel.}
}
\value{
Un `data.frame` con los datos de la tabla `code`.
}
\description{
Conecta a la base de datos DuckDB especificada y recupera la tabla de clasificaciones jerĂ¡rquicas `code` en un R `data.frame`.
Admite tanto el paquete `duckdb` de R como el ejecutable CLI `duckdb` de forma automĂ¡tica.
}
