% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_krona.R
\name{plot_krona}
\alias{plot_krona}
\title{Plot hierarchical Krona chart of taxonomic classifications}
\usage{
plot_krona(df, root_label = "Total", ...)
}
\arguments{
\item{df}{A data.frame of class `btx_code` (from `get_code_table()`) or `btx_cls` (from `classify()`).}

\item{root_label}{Character. Label for the center/root node of the Krona chart. Defaults to `"Total"`.}

\item{...}{Additional arguments passed to `KronaR::kronar_plot`.}
}
\value{
An htmltools tag object (iframe) displaying the interactive Krona chart.
}
\description{
Generates an interactive Krona chart representing the hierarchical structure and
frequency of each taxonomic level from a BacTaxID code table (`btx_code`) or classification results (`btx_cls`).
}
