% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_bactaxid.R
\name{download_bactaxid}
\alias{download_bactaxid}
\title{Download BacTaxID executable}
\usage{
download_bactaxid(
  dest_dir = tools::R_user_dir("bactaxidR", which = "data"),
  force = FALSE
)
}
\arguments{
\item{dest_dir}{Path to the directory where the executable will be saved. Defaults to the user data directory (`tools::R_user_dir("bactaxidR", which = "data")`).}

\item{force}{Logical. If `TRUE`, downloads the executable even if it already exists locally.}
}
\value{
Absolute path to the downloaded executable.
}
\description{
Downloads the precompiled BacTaxID binary (version 1.0 for Linux) from the official GitHub repository.
}
