% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bactaxid.R
\name{download_db}
\alias{download_db}
\title{Descargar base de datos de referencia desde Zenodo}
\usage{
download_db(
  genus,
  dest_dir = tools::R_user_dir("bactaxidR", which = "data"),
  force = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{genus}{Carácter. El género bacteriano de interés (ej. "Escherichia", "Salmonella"). No es sensible a mayúsculas/minúsculas.}

\item{dest_dir}{Ruta al directorio donde se guardará la base de datos descomprimida. Por defecto usa el directorio de datos del usuario (`tools::R_user_dir("bactaxidR", which = "data")`).}

\item{force}{Lógico. Si es TRUE, descarga la base de datos incluso si ya existe localmente.}

\item{verbose}{Lógico. Si es TRUE, muestra mensajes de progreso.}
}
\value{
Ruta absoluta a la base de datos DuckDB descomprimida (`.db`).
}
\description{
Descarga y descomprime la base de datos DuckDB de un género bacteriano específico desde el registro de Zenodo (17791772).
}
