% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bactaxid.R
\name{list_available_dbs}
\alias{list_available_dbs}
\title{Listar bases de datos de géneros disponibles en Zenodo}
\usage{
list_available_dbs(use_api = TRUE)
}
\arguments{
\item{use_api}{Lógico. Si es TRUE, consulta dinámicamente la API de Zenodo. Si es FALSE o si falla
la conexión a la API, utiliza una lista local estática de los 67 géneros disponibles.}
}
\value{
Un vector de caracteres con los nombres de los géneros.
}
\description{
Recupera la lista de géneros bacterianos que tienen archivos de base de datos de referencia
disponibles en el registro de Zenodo de BacTaxID (17791772).
}
