% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sunburst.R
\name{plot_sunburst}
\alias{plot_sunburst}
\title{Plot hierarchical sunburst chart of taxonomic classifications using sunburstR}
\usage{
plot_sunburst(df, root_label = "Total", ...)
}
\arguments{
\item{df}{A data.frame of class `btx_code` (from `get_code_table()`) or `btx_cls` (from `classify()`).}

\item{root_label}{Character. Label for the center/root node of the sunburst chart.
If specified, prepended to all paths. Defaults to `"Total"`.}

\item{...}{Additional arguments passed to `sunburstR::sunburst()`.}
}
\value{
A `sunburst` `htmlwidget` object displaying the interactive sunburst chart.
}
\description{
Generates an interactive sunburst chart representing the hierarchical structure and
frequency of each taxonomic level from a BacTaxID code table (`btx_code`) or classification results (`btx_cls`).
}
